src/backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Bennett-Wendorf <bennettwendorf@gmail.com>"]
license = "MIT"

[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11, <3.12"
quart = "^0.19.3"
torch = {version = "^2.0.1+cpu", source = "torch-cpu"}
scikit-learn = "^1.3.1"
pandas = "^2.1.1"
pathlib = "^1.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
start = "src.api:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
